Moving StatusOr from XLA to stream_executor.
PiperOrigin-RevId: 202179928
